+++ src/app/resource-edit-modal.tsx
@@ -7,6 +7,7 @@
 
 import { useRef, useState } from "react";
 import { type Lang, t } from "./i18n";
+import { ModalFieldError } from "./edit-modal-chrome";
 import { Modal } from "./modal";
 import { ModalHeader } from "./modal-header";
 import { useAutogrow } from "./use-autogrow";
@@ -394,11 +395,7 @@ export function ResourceEditModal({
             </label>
           )}
 
-          {error && (
-            <p role="alert" className="text-sm text-AIPM-pink-strong sm:col-span-2">
-              {error}
-            </p>
-          )}
+          {error && <ModalFieldError error={error} />}
 
           <ModalEditFooter
             lang={lang}